Moved eventlisteners to their on module
